Move package metadata to __about__
This allows `pip install -e .` to access mindsdb version without
pulling in pandas before it's installed.  This works by opening
__about__.py and exec() line by line.

Replace mindsdb.version / mindsdb.version.mindsdb_version with
__about__.py's __version__. Update setup.py to incorporate values
from __about__. This is forward compatible with sphinx's conf.py
should it be used in the future. This way information only needs
to be updated in one place.