Refactor package metadata (#117)
* Move package metadata to __about__

This allows `pip install -e .` to access mindsdb version without
pulling in pandas before it's installed.  This works by opening
__about__.py and exec() line by line.

Replace mindsdb.version / mindsdb.version.mindsdb_version with
__about__.py's __version__. Update setup.py to incorporate values
from __about__. This is forward compatible with sphinx's conf.py
should it be used in the future. This way information only needs
to be updated in one place.

* Allow mindsdb.__version__, pull mindsdb.name from metadata

* Remove flask / flask-socketio dependencies

Package installation seems to stop after:

flask-socketio 3.3.2 has requirement python-socketio<4,>=2.1.0, but
you'll have python-socketio 4.0.0 which is incompatible.

And packages after flask/flask-socketio in requirements are skipped.

Flask/Flask-SocketIO doesn't appear to be used in the package.

* Travis: Install package as editable