Add an error check for \r\n and \n type of strings
problem :
the file_ds.py was not able to find the type of file
given to it due to first_few_lines giving \r\n
so it was not able to find what type of delimiter is used and throws an
error.

Solution :
to fix this an if statement is added to check if line is \r\n or
\n then it will continue else it will append that line in
first_few_lines.