In the updated setup.py file, the dependency_links argument is removed because it is deprecated and not recommended by the latest versions of pip.
Instead of using dependency_links, the URLs of the required packages from private repositories are specified directly in the install_requires list using the git+https syntax.

Therefore, in the install_deps() function, the links list is not needed anymore because it was used to generate the dependency_links argument. The function now only needs to return the new_pkgs list containing the names of the required packages.