Moving MySQL driver to use autocommit
This changes the MySQL integration to use autocommit. This fixes #7234.
The rationale in that ticket is that python errors, if caught, could
lead to dangling open transactions with unexpected results.  This
concern is valid.  If we need transactions we should provide a context
manager for them so that exiting the context manager automaticaly closes
the transaction.

For integration tests I added a test that inserts data, then
disconnects, reconnects, and tries to read it.  This should test data
persistence.