Add custom container-name in example code
Current code is not directly copy-pasteable as-is because "container-name" throws

Error response from daemon: No such container: container-name

Given that the docs are written for complete newbies, it's not clear that "container-name" needs to be replaced with the container name randomly generated by docker run. Instead of instructing the user to find the container name (e.g. with docker ps) it's probably best practice just to set a container name explicitly. Adding that change here with --name