Refactor RAG model in Langchain handler and tools
The code has been refactored to better integrate the RAG (Retrieval-Augmented Generation) model in the Langchain handler. The updates include the addition of tools for creating embedding models, building retrievers and configuring RAG models. A create_retriever_tool function is now used in the handler to get the RAG query tool.