Integrate rag with langchain agents (#9027)
* Add support for RAG model in Langchain handler

This commit introduces integrations for the RAG (Retrieval-Augmented Generation) model in the Langchain handler. It includes various retriever strategies such as SQL, Vector Store, Auto and Multi. Embeddings model support is added and the chunk size for multi-vector retrieval has been increased. A retriever builder is implemented to select the right retriever strategy based on the configuration provided.

* Refactor RAG model in Langchain handler and tools

The code has been refactored to better integrate the RAG (Retrieval-Augmented Generation) model in the Langchain handler. The updates include the addition of tools for creating embedding models, building retrievers and configuring RAG models. A create_retriever_tool function is now used in the handler to get the RAG query tool.

* Refactor Langchain handler to streamline retrieval mode and integrate RAG model building enhancements

* Refactor SkillToolController to use SkillType enum for skill type checks
Refactor SkillToolController to use updated SkillType enum values and improve error messaging for unsupported skill types

* Update method to retrieve RAGPipelineModel fields

The method for retrieving the supported parameters of the RAGPipelineModel is updated from using '__fields__.keys()' to 'schema()['properties'].keys()'. This change ensures the correct keys are returned, aligning the functionality with the updated model schema structure.

* Add support for RAG model in Langchain handler

This commit introduces integrations for the RAG (Retrieval-Augmented Generation) model in the Langchain handler. It includes various retriever strategies such as SQL, Vector Store, Auto and Multi. Embeddings model support is added and the chunk size for multi-vector retrieval has been increased. A retriever builder is implemented to select the right retriever strategy based on the configuration provided.

* Refactor RAG model in Langchain handler and tools

The code has been refactored to better integrate the RAG (Retrieval-Augmented Generation) model in the Langchain handler. The updates include the addition of tools for creating embedding models, building retrievers and configuring RAG models. A create_retriever_tool function is now used in the handler to get the RAG query tool.

* Refactor Langchain handler to streamline retrieval mode and integrate RAG model building enhancements

* Refactor SkillToolController to use SkillType enum for skill type checks
Refactor SkillToolController to use updated SkillType enum values and improve error messaging for unsupported skill types

* Update method to retrieve RAGPipelineModel fields

The method for retrieving the supported parameters of the RAGPipelineModel is updated from using '__fields__.keys()' to 'schema()['properties'].keys()'. This change ensures the correct keys are returned, aligning the functionality with the updated model schema structure.

* Add retrieval model test and update retriever builder

Added a new unit test for retrieval skill in the langchain handler. Made changes in Retriever Builder under integrations/utilities by removing the unsteady SQL Retriever implementation and its related code.

* Enhance vector store handling and make test improvements

Updated handling of vector store instantiation in 'vector_store.py', enabling vector stores to be created from a type and configuration. Made several improvements to 'test_langchain.py' including changes to how the base test class is extended and adjustments to setup and teardown processes. Also, made other minor improvements across various files for better readability and maintainability.

* Update agent creation and retrieval process in Langchain

The commit revises the process of agent creation and information retrieval in the Langchain system. It includes an update to the definition of a retriever skill in `test_langchain.py` and changes the default agent type to `AgentType.OPENAI_FUNCTIONS`. The retrieval process now uses a `RAG` pipeline instead of the previous method and added logging to track pipeline processing. Furthermore, it turned on verbosity in the `langchain_handler`.

* Remove retriever_builder.py file

This commit removes the retriever_builder.py file from the RAG integrations directory within the MindsDB project. The file contained functions for creating and building language retrievers, which are no longer in use.