Fixes for RAG Integration to Python SDK (#9098)
* Allow 'sql' skill name instead of just 'text_to_sql'

* Fixes for RAG integration withj SDK

* Kb update tweaks - DU (#9101)

* Update default params and handle missing embedding args

The default chunk size and overlap values in file_handler have been increased for better performance. In langchain_handler, checks have been added to use the same provider and model for embeddings if 'embedding_model_args' are not provided in input params. Skill_tool's retriever_config has been renamed to 'config' for clarity.

* Handle invoke errors in langchain_handler

This update modifies the langchain handler to accommodate errors during the invocation process. Specifically, it implements exception handling for the agent_executor's invoke method. If an error occurs, instead of crashing, it will now return the error message. However, if the error doesn't match a specific format, the exception will still be raised.

* Update file_handler.py

* Add logging and refactor default vector store in langchain handler

This change introduces logging for situations where 'vector_store_config' is not present in the `langchain_handler` tool configuration. It also modifies the condition that checks for the absence of this property to add a persisting directory. Furthermore, it refactors the code in `langchain_handler.py` related to the absence of 'embedding_model_args'.

* Remove default vector store from langchain handler

The default vector store setup was removed from the langchain handler. The 'vector_store_config' is no longer automatically assigned, and an alert for this missing parameter is no longer logged. The code was adjusted to operate without the previously used 'mindsdb_path'.

* Update warning message format in langchain_handler

Added double quotes around the default collection name in the warning message within langchain_handler. This is to enhance readability and make the default name stand out in the midst of the message.

* Update warning message format in langchain_handler

Added double quotes around the default collection name in the warning message within langchain_handler. This is to enhance readability and make the default name stand out in the midst of the message.

---------

Co-authored-by: Daniel Usvyat <usvyat@gmail.com>