Add table_info method and improve log handling
A new method 'table_info' was added to the 'mindsdb_database_agent' to retrieve all table information in the database. In 'log_callback_handler', the message logging has been changed to iterate through message lists, enhancing detail in debug information. Before it was throwing an error as it was trying to call  pretty_print() on a list