Make agents work with python sdk, support no provider given and just a supported non-minsDB model i.e gpt-4
Renamed multiple instances of ` "embeddings_model"` to `"embedding_model"` for consistency, and removed the `langchain_tools_from_skill` function from the AgentsController and moved it to the LangchainAgent class. Updated the Mindsdb Agent constructor to include the `provider` parameter. Code for model checking and provider validation was refactored and moved to the `check_model_provider` function. An error is now raised in the LangchainAgent if the `prompt` or `llm` components are `None`. Implemented retrieval prompt as the default if none is provided and mode is set to 'retrieval'. List constants in the Agents constant file changed to tuples for immutability.