Refactors CI workflows for improved reliability
Updates CI workflows to use a consistent approach for checking out Github Actions.

This change ensures that the correct version of actions is used, improving the reliability and reproducibility of builds and deployments.

Removes concurrency from the build_deploy_dev workflow as it was causing issues and was not needed, replacing it with concurrency control on the individual jobs.

Moves unit tests into test_on_push and runs them on PRs as this is the expected behaviour.

Adds a check to see if changes are only in docs, and skips tests if so.

Updates deploy workflow to use uv pip install with prerelease enabled.

Switches to using github actions from relative paths.

Adds ENV_URL var to environments.

Fixes json comparison in mysql integration tests.

Removes matrix_includes.json file.

Updates python versions in CI.